requirements.txt
adbc-driver-manager==0.10.0
adbc-driver-postgresql==0.10.0
adbc-driver-sqlite==0.10.0
numpy==1.26.4
polars==0.20.13
pyarrow==15.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
typing_extensions==4.10.0
tzdata==2024.1
